move loadRMGJob functions to tools.loader module
updated sensitivity script
